% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.r
\name{est.all.wb}
\alias{est.all.wb}
\title{Get stage 1 & 2 estimates and T and B cell subtypes}
\usage{
est.all.wb(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A numeric matrix
}
\description{
Get stage 1 & 2 estimates and T and B cell subtypes
}

