% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_functions.r
\name{fit.cd19.sorted}
\alias{fit.cd19.sorted}
\title{Fit model for CD19 subtypes using M450 data from sorted B cells}
\usage{
fit.cd19.sorted(data, cpg.list = cpg.list.cd19)
}
\arguments{
\item{data}{A data frame}

\item{cpg.list}{A character}
}
\value{
A numeric matrix
}
\description{
Fit model for CD19 subtypes using M450 data from sorted B cells
}
\seealso{
Other Model fitting: \code{\link{MethylDeconBloodSubtypes}},
  \code{\link{MethylDeconBloodSubtypes-package}},
  \code{\link{fitNonneg.nosum1}};
  \code{\link{fit.cd19.sub}}; \code{\link{fit.cd4.sorted}};
  \code{\link{fit.cd8.sub.sorted}};
  \code{\link{fit.cd8.sub}}; \code{\link{fit.main.nosum1}};
  \code{\link{fit.stage2b}};
  \code{\link{get.cpg.list.cd19.sub}}
}

