% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.r
\name{fit.cd4.sorted}
\alias{fit.cd4.sorted}
\title{Fit model for CD4 subtypes on M450 data from sorted CD4 T cells}
\usage{
fit.cd4.sorted(data, cpg.list = cpg.list.cd4)
}
\arguments{
\item{data}{A data frame}

\item{cpg.list}{A character}
}
\value{
A numeric matrix
}
\description{
Fit model for CD4 subtypes on M450 data from sorted CD4 T cells
}
\seealso{
Other Model fitting: \code{\link{MethylDeconBloodSubtypes}},
  \code{\link{fit.cd19.sorted}},
  \code{\link{fit.cd19.sub}},
  \code{\link{fit.cd8.sub.sorted}},
  \code{\link{fit.cd8.sub}}, \code{\link{fit.main.nosum1}},
  \code{\link{fit.stage2b}},
  \code{\link{get.cpg.list.cd19.sub}}
}

