% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.r
\name{fit.cd4.sub}
\alias{fit.cd4.sub}
\title{Fit model for CD4 subtypes on whole blood data}
\usage{
fit.cd4.sub(data, fit.main, cpg.list = cpg.list.cd4)
}
\arguments{
\item{data}{A data frame}

\item{fit.main}{A numeric matrix}

\item{cpg.list}{A character}
}
\value{
A numeric matrix
}
\description{
Fit model for CD4 subtypes on whole blood data
}
\seealso{
Other Generate list: \code{\link{get.cpg.list.cd4.sub}},
  \code{\link{get.cpg.list.cd8.sub}},
  \code{\link{get.cpg.list.main}},
  \code{\link{get.cpg.list.stage2b}}
}

