% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_functions.r
\name{fit.cd8.sub}
\alias{fit.cd8.sub}
\title{Fit model for CD8 subtypes for whole blood data}
\usage{
fit.cd8.sub(data, fit.main, cpg.list = cpg.list.cd8)
}
\arguments{
\item{data}{A data frame}

\item{fit.main}{A numeric matrix}

\item{cpg.list}{A character}
}
\value{
A numeric matrix
}
\description{
Fit model for CD8 subtypes for whole blood data
}
\seealso{
Other Model fitting: \code{\link{MethylDeconBloodSubtypes}},
  \code{\link{MethylDeconBloodSubtypes-package}},
  \code{\link{fitNonneg.nosum1}};
  \code{\link{fit.cd19.sorted}};
  \code{\link{fit.cd19.sub}}; \code{\link{fit.cd4.sorted}};
  \code{\link{fit.cd8.sub.sorted}};
  \code{\link{fit.main.nosum1}}; \code{\link{fit.stage2b}};
  \code{\link{get.cpg.list.cd19.sub}}
}

