% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_functions.r
\name{fit.main.nosum1}
\alias{fit.main.nosum1}
\title{Fit the model for the 6 main cell types}
\usage{
fit.main.nosum1(data, cpg.list = cpg.list.main.stage1)
}
\arguments{
\item{data}{A data frame}

\item{cpg.list}{A character}
}
\value{
A character
}
\description{
Fit the model for the 6 main cell types
}
\seealso{
Other Model fitting: \code{\link{MethylDeconBloodSubtypes}},
  \code{\link{MethylDeconBloodSubtypes-package}},
  \code{\link{fitNonneg.nosum1}};
  \code{\link{fit.cd19.sorted}};
  \code{\link{fit.cd19.sub}}; \code{\link{fit.cd4.sorted}};
  \code{\link{fit.cd8.sub.sorted}};
  \code{\link{fit.cd8.sub}}; \code{\link{fit.stage2b}};
  \code{\link{get.cpg.list.cd19.sub}}
}

