% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.r
\name{fit.stage2b}
\alias{fit.stage2b}
\title{Fit models for stage 2}
\usage{
fit.stage2b(data, fit.main, cpg.list = cpg.list.2b)
}
\arguments{
\item{data}{A data frame}

\item{fit.main}{A numeric matrix}

\item{cpg.list}{A character}
}
\value{
A numeric matrix
}
\description{
Fit models for stage 2
}
\seealso{
Other Model fitting: \code{\link{MethylDeconBloodSubtypes}},
  \code{\link{fit.cd19.sorted}},
  \code{\link{fit.cd19.sub}}, \code{\link{fit.cd4.sorted}},
  \code{\link{fit.cd8.sub.sorted}},
  \code{\link{fit.cd8.sub}}, \code{\link{fit.main.nosum1}},
  \code{\link{get.cpg.list.cd19.sub}}
}

