% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.r
\name{get.beta.sub}
\alias{get.beta.sub}
\title{Partition whole blood beta for two-stage model fit}
\usage{
get.beta.sub(data, blood = med.main, fit, keep.cols)
}
\arguments{
\item{data}{A data frame}
}
\value{
A numeric matrix
}
\description{
Partition whole blood beta for two-stage model fit
}
\seealso{
Other Data partitioning: \code{\link{get.beta.sub.inner}}
}

