% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.r
\name{get.cpg.list.cd19.sub}
\alias{get.cpg.list.cd19.sub}
\title{Function to get CpG list for CD19 B cell subtypes}
\usage{
get.cpg.list.cd19.sub(length.pairs = 300)
}
\arguments{
\item{length.pairs}{An integer numeric}
}
\value{
A character
}
\description{
Function to get CpG list for CD19 B cell subtypes
}
\seealso{
Other Model fitting: \code{\link{MethylDeconBloodSubtypes}},
  \code{\link{fit.cd19.sorted}},
  \code{\link{fit.cd19.sub}}, \code{\link{fit.cd4.sorted}},
  \code{\link{fit.cd8.sub.sorted}},
  \code{\link{fit.cd8.sub}}, \code{\link{fit.main.nosum1}},
  \code{\link{fit.stage2b}}
}

