% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_functions.r
\name{get.cpg.list.cd4.sub}
\alias{get.cpg.list.cd4.sub}
\title{Get CpG list for CD4 subtypes}
\usage{
get.cpg.list.cd4.sub(length.anova = 2100, length.pairs = 600)
}
\arguments{
\item{length.anova}{An integer numeric}

\item{length.pairs}{An integer numeric}
}
\value{
A character
}
\description{
Get CpG list for CD4 subtypes
}
\seealso{
Other Generate list: \code{\link{fit.cd4.sub}};
  \code{\link{get.cpg.list.cd8.sub}};
  \code{\link{get.cpg.list.main}};
  \code{\link{get.cpg.list.stage2b}}
}

