% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.r
\name{get.cpg.list.cd8.sub}
\alias{get.cpg.list.cd8.sub}
\title{Get CpG list for CD8 subtypes}
\usage{
get.cpg.list.cd8.sub(length.list = 300)
}
\arguments{
\item{length.list}{An integer numeric}
}
\value{
A character
}
\description{
Get CpG list for CD8 subtypes
}
\seealso{
Other Generate list: \code{\link{fit.cd4.sub}},
  \code{\link{get.cpg.list.cd4.sub}},
  \code{\link{get.cpg.list.main}},
  \code{\link{get.cpg.list.stage2b}}
}

