% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_functions.r
\name{get.cpg.list.main}
\alias{get.cpg.list.main}
\title{Get the CpG list for the model for main cell types.}
\usage{
get.cpg.list.main(length.anova = 1000, length.pairs = 300)
}
\arguments{
\item{length.anova}{An integer numeric}

\item{length.pairs}{An integer numeric}
}
\value{
A character
}
\description{
Get the CpG list for the model for main cell types.
}
\seealso{
Other Generate list: \code{\link{fit.cd4.sub}};
  \code{\link{get.cpg.list.cd4.sub}};
  \code{\link{get.cpg.list.cd8.sub}};
  \code{\link{get.cpg.list.stage2b}}
}

