% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.r
\name{get.cpg.list.stage2b}
\alias{get.cpg.list.stage2b}
\title{Get CpG list for stage 2}
\usage{
get.cpg.list.stage2b(length.pairs = 100)
}
\arguments{
\item{length.pairs}{An integer numeric}
}
\value{
A Character
}
\description{
Get CpG list for stage 2
}
\seealso{
Other Generate list: \code{\link{fit.cd4.sub}},
  \code{\link{get.cpg.list.cd4.sub}},
  \code{\link{get.cpg.list.cd8.sub}},
  \code{\link{get.cpg.list.main}}
}

