% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_functions.r
\docType{package}
\name{MethylDeconBloodSubtypes}
\alias{MethylDeconBloodSubtypes}
\alias{MethylDeconBloodSubtypes-package}
\alias{fitNonneg.nosum1}
\title{Deconvolute array-based methylation data.
Fit models.}
\usage{
fitNonneg.nosum1(testVector, basis)
}
\arguments{
\item{testVector}{}

\item{basis}{}
}
\value{
A numeric matrix
}
\description{
Deconvolute array-based methylation data.
Fit models.
}
\seealso{
Other Model fitting: \code{\link{fit.cd19.sorted}};
  \code{\link{fit.cd19.sub}}; \code{\link{fit.cd4.sorted}};
  \code{\link{fit.cd8.sub.sorted}};
  \code{\link{fit.cd8.sub}}; \code{\link{fit.main.nosum1}};
  \code{\link{fit.stage2b}};
  \code{\link{get.cpg.list.cd19.sub}}
}

