% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.r
\name{get.beta.sub.inner}
\alias{get.beta.sub.inner}
\title{Partition whole blood beta for two-stage model fit inner}
\usage{
get.beta.sub.inner(x, fit, keep.cols)
}
\value{
A numeric matrix
}
\description{
Partition whole blood beta for two-stage model fit inner
}
\seealso{
Other Data partitioning: \code{\link{get.beta.sub}}
}

